+++ src/index.ts
@@ -1,6 +1,7 @@
 import { dl } from './dl'
 import { Vampire } from './vampire'
 import { readUrl } from './read-url'
+import { is404Error } from './util'
 
 // types
 export type { DlOptions } from './dl'
@@ -12,7 +13,38 @@ export type {
   Progress,
 } from './vampire'
 
-export { Vampire, readUrl, dl }
+import {
+  // all got.*Error
+  RequestError,
+  HTTPError,
+  MaxRedirectsError,
+  UnsupportedProtocolError,
+  TimeoutError,
+  ReadError,
+  CacheError,
+  ParseError,
+  CancelError,
+  UploadError,
+} from 'got'
+
+export {
+  Vampire,
+  readUrl,
+  dl,
+  is404Error,
+
+  // got
+  RequestError,
+  HTTPError,
+  MaxRedirectsError,
+  UnsupportedProtocolError,
+  TimeoutError,
+  ReadError,
+  CacheError,
+  ParseError,
+  CancelError,
+  UploadError,
+}
 export default dl
 
 // https://github.com/sindresorhus/got/blob/v11.8.3/source/index.ts#L127
@@ -22,4 +54,17 @@ Object.assign(module.exports, {
   default: dl,
   Vampire,
   readUrl,
+  is404Error,
+
+  // errors
+  RequestError,
+  HTTPError,
+  MaxRedirectsError,
+  UnsupportedProtocolError,
+  TimeoutError,
+  ReadError,
+  CacheError,
+  ParseError,
+  CancelError,
+  UploadError,
 })